Final version

+++ Bmi.h
@@ -6,9 +6,9 @@ public:
 	Bmi();
 	Bmi(int m, int h);
 	void setValue(int m,int h);
-	float getValue();
+	double getValue();
 	int getCate(double);	
 private:
 	int mass, height;
-	float result;
+	double result;
 };
